一些批量脚本和实用工具/py3-phpmyadmin-less4.6/requirements.txt
requests==2.19.1
colorama==0.3.9
progressbar2==3.38.0
